+++ udata_tabular_preview/__init__.py
@@ -4,5 +4,5 @@ Tabular Preview
 A plugin to preview tabular files (CSV, Excel, ...)
 '''
 
-__version__ = '3.0.3'
+__version__ = '3.0.4.dev'
 __description__ = 'A plugin to preview tabular files (CSV, Excel, ...) using csvapi'