+++ tasks.py
@@ -1,8 +1,17 @@
 import os
 
+from datetime import datetime
+
+from babel.messages.pofile import read_po, write_po
+from babel.util import LOCALTZ
+
 from invoke import task, call
 
 ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__)))
+
+PYTHON_I18N_ROOT = 'udata_tabular_preview/translations'
+
+LANGUAGES = ['fr']
 
 TO_CLEAN = ['build', 'dist', '**/*.pyc', 'reports']
 
@@ -111,10 +120,39 @@ def assets_build(ctx):
         ctx.run('npm run build', pty=True)
 
 
+def set_po_metadata(filename, locale):
+    # Fix crowdin requiring Language with `2-digit` iso code in potfile
+    # to produce 2-digit iso code pofile
+    # Opening the catalog also allows to set extra metadata
+    with open(filename, 'rb') as infile:
+        catalog = read_po(infile, locale)
+    catalog.copyright_holder = 'Etalab'
+    catalog.msgid_bugs_address = 'data.gouv@data.gouv.fr'
+    catalog.language_team = 'Data.gouv.fr Team <data.gouv@data.gouv.fr>'
+    catalog.last_translator = 'Data.gouv.fr Team <data.gouv@data.gouv.fr>'
+    catalog.revision_date = datetime.now(LOCALTZ)
+    with open(filename, 'wb') as outfile:
+        write_po(outfile, catalog, width=80)
+
 @task
 def i18n(ctx, update=False):
     '''Extract translatable strings'''
     header(i18n.__doc__)
+
+    # Python translations
+    info('Extract python translations')
+    with ctx.cd(ROOT):
+        ctx.run('python setup.py extract_messages')
+        set_po_metadata(os.path.join(PYTHON_I18N_ROOT, 'udata_tabular_preview.pot'), 'en')
+        for lang in LANGUAGES:
+            pofile = os.path.join(PYTHON_I18N_ROOT, lang, 'LC_MESSAGES', 'udata_tabular_preview.po')
+            if not os.path.exists(pofile):
+                ctx.run('python setup.py init_catalog -l {}'.format(lang))
+                set_po_metadata(pofile, lang)
+            elif update:
+                ctx.run('python setup.py update_catalog -l {}'.format(lang))
+                set_po_metadata(pofile, lang)
+
 
     # Front translations
     info('Extract vue translations')
@@ -123,7 +161,15 @@ def i18n(ctx, update=False):
     success('Updated translations')
 
 
-@task(assets_build)
+@task
+def i18nc(ctx):
+    '''Compile translations'''
+    header('Compiling translations')
+    with ctx.cd(ROOT):
+        ctx.run('python setup.py compile_catalog')
+
+
+@task(i18nc, assets_build)
 def dist(ctx, buildno=None):
     '''Package for distribution'''
     header('Building a distribuable package')
@@ -135,7 +181,7 @@ def dist(ctx, buildno=None):
         ctx.run(' '.join(cmd), pty=True)
     success('Distribution is available in dist directory')
 
-@task
+@task(i18nc)
 def pydist(ctx, buildno=None):
     '''Perform python packaging (without compiling assets)'''
     header('Building a distribuable package')